check password upon writing / deleting content

+++ srv/routes/projects.js
@@ -19,7 +19,6 @@ router.get('/', function(req, res, next) {
 router.get('/edit/:id', function(req, res, next) {
 	const id = req.params.id;
 	const sql = "SELECT * FROM projects WHERE id = ?";
-	// '?' will be transformed to SQL injection safe id in db.get()
 	db.get(sql, id, (err, row) => {
 		if (err) return console.error(err.message);
 		res.render("projects/edit", { model: row });
@@ -27,6 +26,12 @@ router.get('/edit/:id', function(req, res, next) {
 });
 
 router.post("/edit/:id", (req, res) => {
+	if (req.body.password != process.env.APP_PASSWD) {
+		res.status(403);
+		req.body.error = "bad password";
+		res.render("projects/edit", { model: req.body });
+		return ;
+	}
 	const id = req.params.id;
 	// TODO upload image
 	const project = [
@@ -54,6 +59,12 @@ router.get("/create", (req, res) => {
 
 router.post("/create", (req, res) => {
 	// TODO upload image
+	if (req.body.password != process.env.APP_PASSWD) {
+		res.status(403);
+		req.body.error = "bad password";
+		res.render("projects/create", { model: req.body });
+		return ;
+	}
 	const project = [
 		req.body.title,
 		req.body.type,
@@ -74,10 +85,30 @@ router.post("/create", (req, res) => {
 
 router.get("/delete/:id", (req, res) => {
 	const id = req.params.id;
-	const sql = "DELETE FROM projects WHERE id = ?";
-	db.run(sql, id, err => {
+	const sql = "SELECT * FROM projects WHERE id = ?";
+	db.get(sql, id, (err, row) => {
 		if (err) return console.error(err.message);
-		res.redirect("/projects");
+		res.render("projects/delete", { model: row });
+	});
+});
+
+router.post("/delete/:id", (req, res) => {
+	const id = req.params.id;
+	const sql = "SELECT * FROM projects WHERE id = ?";
+	db.get(sql, id, (err, row) => {
+		if (err) return console.error(err.message);
+		if (req.body.password != process.env.APP_PASSWD) {
+			res.status(403);
+			row.error = "bad password";
+			res.render("projects/delete", { model: row});
+			return ;
+		}
+		const id = req.params.id;
+		const sql = "DELETE FROM projects WHERE id = ?";
+		db.run(sql, id, err => {
+			if (err) return console.error(err.message);
+			res.redirect("/projects");
+		});
 	});
 });
